requirements.txt
asgiref==3.5.0
beautifulsoup4==4.11.1
certifi==2021.5.30
dj-database-url==0.5.0
Django==4.0.4
django-ckeditor==6.4.0
django-crispy-forms==1.14.0
django-etc==1.3.1
django-hitcount==1.3.5
django-js-asset==2.0.0
django-pygmentify==0.3.7
djangorestframework==3.13.1
Pillow==9.1.0
Pygments==2.12.0
pytz==2022.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
whitenoise==6.1.0
gunicorn==20.1.0
psycopg2-binary==2.9.3